requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudscraper==1.2.71
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
oauthlib==3.2.2
openai==1.63.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlalchemy==2.0.28
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
